Boilerplate for chord midifx. Moved all the structs used for chords to it's own header. Added chordUtil class that can be used for things done in both chord mode and chord midifx.
