+++ OMX-27-firmware/src/utils/chord_util.cpp
@@ -1,0 +1,12 @@
+#include <U8g2_for_Adafruit_GFX.h>
+
+#include "chord_util.h"
+#include "../consts/consts.h"
+#include "../midi/midi.h"
+#include "../consts/colors.h"
+#include "../hardware/omx_leds.h"
+#include "../hardware/omx_disp.h"
+#include "../midi/noteoffs.h"
+#include "../modes/sequencer.h"
+
+ChordUtil chordUtil;
